requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
discord.py==2.3.2
dnspython==2.4.2
easyocr==1.7.1
filelock==3.12.4
frozenlist==1.4.0
idna==3.4
imageio==2.31.4
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
ninja==1.11.1
numpy==1.26.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
pyclipper==1.3.0.post5
pymongo==4.5.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.2
shapely==2.0.1
six==1.16.0
sympy==1.12
tifffile==2023.9.18
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
yarl==1.9.2
